+++ app/[locale]/projects/[projectId]/dmp/[dmpid]/related-works/actions/index.ts
@@ -1,1 +1,0 @@
-export { updateRelatedWorkStatusAction } from './updateRelatedWorkStatusAction';