+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 - Added new queries `GuidanceSourcesForPlan`, `AddPlanGuidance`, `RemovePlanGuidance` and `ManagedAffiliationsWithGuidance` [#15]
 - Added new component `AffiliationSearchForGuidance` to be used for the custom guidance search modal [#15]
 - Added a new type policy `guidanceTypePolicies` for Apollo Client [#15]
+- Added related works project overview page [#700]
+
+## Updated
+
+
+## Chore
+- Updated version of `next` to `v16.1.5` due to vulnerability and `next-intl` to `4.8.0` due to dependency
+- Updated version of `@dmptool/types` to `v3.1.1` because it has a fix for the 'fs' errors
+==========================================================================================
+## All changes above the line happened after the merge to the main branch on Jan 27, 2026
+==========================================================================================
+## Added
 - Added new `GuidancePanel` component for the tabbed guidance in the right sidebar [#12]
 - Added `BestPracticeGuidance` query [#12]
 - Added ability to show if there are pending related works and the number of accepted works [#981]